Separando a classe dos planetas
