+++ examples/sistemaSolar/CMakeLists.txt
@@ -1,3 +1,3 @@
 project(sistemaSolar)
-add_executable(${PROJECT_NAME} camera.cpp main.cpp window.cpp)
+add_executable(${PROJECT_NAME} planet.cpp camera.cpp main.cpp window.cpp)
 enable_abcg(${PROJECT_NAME})
