+++ examples/sistemaSolar/planet.cpp
@@ -1,0 +1,141 @@
+#include "planet.hpp"
+
+#include <glm/gtx/fast_trigonometry.hpp>
+#include <unordered_map>
+
+// Explicit specialization of std::hash for Vertex
+template <> struct std::hash<Vertex> {
+  size_t operator()(Vertex const &vertex) const noexcept {
+    auto const h1{std::hash<glm::vec3>()(vertex.position)};
+    return h1;
+  }
+};
+
+void Planet::create(GLuint program, std::string assetsPath) {
+  destroy();
+
+  m_program = program;
+
+  m_colorLoc = abcg::glGetUniformLocation(m_program, "color");
+  m_scaleLoc = abcg::glGetUniformLocation(m_program, "scale");
+  m_translationLoc = abcg::glGetUniformLocation(m_program, "translation");
+  m_modelMatrixLoc = abcg::glGetUniformLocation(m_program, "modelMatrix");
+
+  m_translation = glm::vec3(7.0f, 0.0f, 0.0f);
+  m_color = glm::vec4(0.9f, 0.9f, 0.9f, 1.0f);
+  m_velocity = glm::vec3(0, 0, 0);
+
+  // Load model
+  loadModelFromFile(assetsPath + "sphere.obj");
+
+  // Generate VBO
+  abcg::glGenBuffers(1, &m_VBO);
+  abcg::glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
+  abcg::glBufferData(GL_ARRAY_BUFFER,
+                     sizeof(m_vertices.at(0)) * m_vertices.size(),
+                     m_vertices.data(), GL_STATIC_DRAW);
+  abcg::glBindBuffer(GL_ARRAY_BUFFER, 0);
+
+  // Generate EBO
+  abcg::glGenBuffers(1, &m_EBO);
+  abcg::glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_EBO);
+  abcg::glBufferData(GL_ELEMENT_ARRAY_BUFFER,
+                     sizeof(m_indices.at(0)) * m_indices.size(),
+                     m_indices.data(), GL_STATIC_DRAW);
+  abcg::glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
+
+  // Create VAO
+  abcg::glGenVertexArrays(1, &m_VAO);
+
+  // Bind vertex attributes to current VAO
+  abcg::glBindVertexArray(m_VAO);
+
+  abcg::glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
+  auto const positionAttribute{
+      abcg::glGetAttribLocation(m_program, "inPosition")};
+  abcg::glEnableVertexAttribArray(positionAttribute);
+  abcg::glVertexAttribPointer(positionAttribute, 3, GL_FLOAT, GL_FALSE,
+                              sizeof(Vertex), nullptr);
+  abcg::glBindBuffer(GL_ARRAY_BUFFER, 0);
+
+  abcg::glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_EBO);
+
+  // End of binding to current VAO
+  abcg::glBindVertexArray(0);
+}
+
+void Planet::loadModelFromFile(std::string_view path) {
+  tinyobj::ObjReader reader;
+
+  if (!reader.ParseFromFile(path.data())) {
+    if (!reader.Error().empty()) {
+      throw abcg::RuntimeError(
+          fmt::format("Failed to load model {} ({})", path, reader.Error()));
+    }
+    throw abcg::RuntimeError(fmt::format("Failed to load model {}", path));
+  }
+
+  if (!reader.Warning().empty()) {
+    fmt::print("Warning: {}\n", reader.Warning());
+  }
+
+  auto const &attributes{reader.GetAttrib()};
+  auto const &shapes{reader.GetShapes()};
+
+  m_vertices.clear();
+  m_indices.clear();
+
+  // A key:value map with key=Vertex and value=index
+  std::unordered_map<Vertex, GLuint> hash{};
+
+  // Loop over shapes
+  for (auto const &shape : shapes) {
+    // Loop over indices
+    for (auto const offset : iter::range(shape.mesh.indices.size())) {
+      // Access to vertex
+      auto const index{shape.mesh.indices.at(offset)};
+
+      // Vertex position
+      auto const startIndex{3 * index.vertex_index};
+      auto const vx{attributes.vertices.at(startIndex + 0)};
+      auto const vy{attributes.vertices.at(startIndex + 1)};
+      auto const vz{attributes.vertices.at(startIndex + 2)};
+
+      Vertex const vertex{.position = {vx, vy, vz}};
+
+      // If map doesn't contain this vertex
+      if (!hash.contains(vertex)) {
+        // Add this index (size of m_vertices)
+        hash[vertex] = m_vertices.size();
+        // Add this vertex
+        m_vertices.push_back(vertex);
+      }
+
+      m_indices.push_back(hash[vertex]);
+    }
+  }
+}
+
+void Planet::paint() {
+  abcg::glUseProgram(m_program);
+
+  abcg::glBindVertexArray(m_VAO);
+
+  glm::mat4 model{1.0f};
+  model = glm::mat4(1.0);
+  model = glm::translate(model, glm::vec3(0.0f, 0.0f, 0.0f));
+  model = glm::scale(model, glm::vec3(0.4f));
+
+  abcg::glUniformMatrix4fv(m_modelMatrixLoc, 1, GL_FALSE, &model[0][0]);
+  abcg::glUniform4f(m_colorLoc, 1.0f, 0.8f, 0.0f, 1.0f);
+  abcg::glDrawElements(GL_TRIANGLES, m_indices.size(), GL_UNSIGNED_INT,
+                       nullptr);
+
+  abcg::glBindVertexArray(0);
+
+  abcg::glUseProgram(0);
+}
+
+void Planet::destroy() {}
+
+void Planet::update() {}
